backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
pydantic==2.12.5
pydantic-settings==2.11.0
email-validator==2.3.0
bcrypt==4.1.3
PyJWT==2.12.1
python-multipart==0.0.22
APScheduler==3.11.2
sendgrid==6.12.5
